Create README.md
ECOMVISON Admin Dashboard
The ECOMVISON Admin Dashboard is a full-stack web application designed for portfolio purposes. It serves as an admin dashboard to manage various aspects of an e-commerce platform, including products, customers, transactions, sales data, and user performance. The dashboard is built using the MERN stack (MongoDB, Express, React, Node.js) and features role-based access control, allowing different roles to have specific permissions and restrictions.

While the primary focus of the project was to showcase development skills, please note that the mobile version of the dashboard may not be fully optimized for responsiveness on all devices.

Explore the live demo of the ECOMVISON Admin Dashboard on Render.com and feel free to provide feedback or reach out with any inquiries at oussamaqqqq@gmail.com